# openapi.yaml
openapi: 3.0.3
info:
  title: KOI API
  description: |
    REST API for **KOI (Kubernetes on Incus)** — see [`API.md`](API.md) at the
    repo root for the full prose reference this spec is generated to match
    (async job pattern, error shape, per-owner scoping rules, size-string
    format, worked examples). This file is the machine-readable contract;
    `API.md` is the human-readable one — keep both in sync.

    ### Auth
    Session is a JWT in an **HttpOnly cookie** (`auth_token`), set by
    `/auth/register-admin` and `/auth/login`. There is no bearer-token
    header — browser clients just need `credentials: "include"`; non-browser
    clients must persist and resend the `Set-Cookie` value.

    ### Scoping
    Every Cluster Network / Cluster / Node / Job is scoped to its owner
    (the logged-in user). A resource owned by someone else looks identical
    to a nonexistent one (`404`), never a `403` — this avoids leaking
    existence. The one exception is the `/users` endpoints, which are
    admin-only and return `403` for a logged-in non-admin.

    ### The async job pattern
    `POST /clusters`, `POST /clusters/{id}/nodes`, `DELETE
    /clusters/{id}/nodes/{nodeId}`, and `DELETE /clusters/{id}` all return
    `202` immediately with the affected resource in a non-final state
    (`creating`/`deleting`) and a `jobId` to poll via `GET /jobs/{id}` until
    its `status` becomes `succeeded` or `failed`.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/auth/register-admin:
    post:
      tags: [Auth]
      operationId: registerAdmin
      summary: Bootstrap the one admin account (works exactly once)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialsRequest" }
      responses:
        "201":
          description: Admin created and logged in (session cookie set)
          headers:
            Set-Cookie: { $ref: "#/components/headers/SessionCookie" }
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: An admin already exists — use /auth/login instead
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/users:
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a regular user
      description: Admin-only. `role` is always `"user"` — the one admin account only ever comes from `/auth/register-admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialsRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Username already taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users
      description: Admin-only. Newest first.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Users]
      operationId: getUser
      summary: Get a single user
      description: Admin-only.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete a non-admin user and everything they own
      description: |
        Admin-only, and only for non-admin targets — since the caller must
        already be an admin, this also means an admin can never delete
        themselves or another admin here. Starts a background job
        (`type: "user_deletion"`, owned by the admin who triggered it) that
        deletes every VM in every cluster the target owns, those clusters,
        every network they own, and finally the user account itself.
      responses:
        "202":
          description: Accepted — deletion running in the background
          content:
            application/json:
              schema:
                type: object
                properties:
                  job: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Not an admin, or the target user is an admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --------------------------------------------------------------- Networks
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HttpOnly JWT session cookie, set by /auth/register-admin or /auth/login.
  headers:
    SessionCookie:
      description: Sets or clears the `auth_token` HttpOnly cookie.
      schema: { type: string }
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: User ID.
...
  responses:
    BadRequest:
      description: Validation error — safe to show `message` next to the offending field
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Not logged in, or session cookie missing/invalid/expired
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Logged in, but this account's role isn't allowed to do this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found (or belongs to someone else — the two look identical, by design)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      description: Shape of every non-2xx response.
      required: [error, message, code]
      properties:
        error:
          type: string
          description: Short machine-ish category. Not a stable enum — switch on the HTTP status code instead.
          example: validation error
        message:
          type: string
          description: Human-readable detail, safe to show directly in a UI.
          example: "cpu must be at least 2, got 1"
        code:
          type: integer
          description: Repeats the HTTP status code.
          example: 400
    CredentialsRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 63
        password:
          type: string
          minLength: 8
          description: Minimum 8 characters, enforced for every account (admin or regular user).
    User:
      type: object
      description: Never includes a password/hash — excluded from JSON entirely, server-side.
      required: [id, username, role, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        role: { type: string, enum: [admin, user] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Job:
      type: object
      required: [id, ownerId, type, status, progress, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        type:
          type: string
          enum: [node_provision, node_deletion, cluster_deletion, user_deletion]
          description: Check metadata.role (node_provision) / metadata.nodeId / metadata.clusterId / metadata.userId to know which resource this job is about.
        name: { type: string }
        status: { type: string, enum: [queued, running, succeeded, failed] }
        progress: { type: integer, minimum: 0, maximum: 100 }
        stage:
          type: string
          description: Informational progress string (e.g. "launching", "bootstrapping", "joining") — not a fixed enum, just display it.
        message: { type: string }
        result:
          type: object
          additionalProperties: true
          nullable: true
          description: 'Populated on success. Currently just {"ip": "..."} for node_provision.'
        error:
          type: string
          nullable: true
          description: Raw underlying failure detail (often multi-line command output) — fine for a collapsible "details" section, not the headline message.
        metadata:
          type: object
          additionalProperties: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }